- Added basic translation gizmo

+++ Apps/Editor/src/views/AssetViewer/AssetViewer.cpp
@@ -30,10 +30,10 @@ namespace Osiris::Editor
 
 	struct LayoutSettings_s
 	{
-		int itemGroupWidth = 64;
+		float itemGroupWidth = 64.0f;
 		int itemColumnCnt = 4;
-		int itemGroupPaddingX = 6;
-		int itemColumnWidth;
+		float itemGroupPaddingX = 6.0f;
+		float itemColumnWidth;
 		int itemLabelShortCharLimit = 9;
 	} layoutSettings;
 
@@ -500,7 +500,7 @@ namespace Osiris::Editor
 		ImGui::SameLine();
 		ImGui::SetCursorPosX((ImGui::GetCursorPosX() - groupSize.x) + layoutSettings.itemGroupPaddingX);
 
-		ImGui::Image((ImTextureID)(INT_PTR)textureResource->GetTexture()->GetHandle(), ImVec2(std::clamp(textureResource->GetWidth(), (uint32_t)1, (uint32_t)layoutSettings.itemGroupWidth), std::clamp(textureResource->GetHeight(), (uint32_t)1, (uint32_t)layoutSettings.itemGroupWidth)));
+		ImGui::Image((ImTextureID)(INT_PTR)textureResource->GetTexture()->GetHandle(), ImVec2((float)std::clamp(textureResource->GetWidth(), (uint32_t)1, (uint32_t)layoutSettings.itemGroupWidth), (float)std::clamp(textureResource->GetHeight(), (uint32_t)1, (uint32_t)layoutSettings.itemGroupWidth)));
 		ImGui::SetCursorPos(ImVec2(ImGui::GetCursorPosX() + 2, ImGui::GetCursorPosY() - labelSize.y));
 
 		if (labelSize.x > (layoutSettings.itemColumnWidth - 4))
